requirements.txt
flask
openai
python-dotenv
requests
beautifulsoup4
lxml
pandas
gunicorn
